restructured the styles components
