+++ src/appTheme/AppTheme.tsx
@@ -2,9 +2,7 @@ import { StyleSheet } from 'react-native';
 import { flightInfoHeight, rest } from '../helpers/DataDestinyDimensions';
 
 
-
 const styles = StyleSheet.create({
-  // SIGN UP STYLES
   container: {
     flex: 0,
     marginTop: 30,
@@ -59,236 +57,6 @@ const styles = StyleSheet.create({
     marginLeft: 10,
   },
 
-  // BUTTON PRIMARY STYLES
-  buttonContainer: {
-    alignItems: 'center',
-    alignSelf: 'center',
-    borderRadius: 10,
-    marginTop: 10,
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-    width: '95%',
-    height: 50,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 4,
-    },
-    shadowOpacity: 0.3,
-    shadowRadius: 3,
-    elevation: 4,
-  },
-  primaryButtonValid: {
-    backgroundColor: '#2774D5',
-  },
-  primaryButtonInvalid: {
-    backgroundColor: '#A0A0A0',
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-  buttonImage: {
-    position: 'absolute',
-    left: '10%',
-    width: 25,
-    height: 25,
-  },
-
-  //CUSTOM CHECKBOX STYLES
-  checkButton: {
-    borderWidth: 1,
-    width: 18,
-    height: 18,
-    marginRight: '2.5%',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  checkbox: {
-    flexDirection: 'row',
-    marginBottom: 5,
-  },
-  checkboxSelected: {
-    backgroundColor: '#2D7BDD',
-  },
-  checkIcon: {
-    width: 12,
-    height: 12,
-    tintColor: 'white',
-  },
-  checkboxUnselected: {
-    backgroundColor: 'transparent',
-  },
-  checkboxContainer: {
-    flexDirection: 'column',
-    marginBottom: 10,
-  },
-  errorCheckBox: {
-    color: 'red',
-    fontSize: 13,
-  },
-
-  //INPUTFIELD STYLES
-  inputContainer: {
-    marginBottom: 10,
-    alignItems: 'center',
-    justifyContent: 'flex-start',
-  },
-  inputLabelContainer: {
-    flexDirection: 'row',
-    alignSelf: 'flex-start',
-  },
-  inputLabel: {
-    marginBottom: 5,
-    fontSize: 15,
-    fontWeight: 'bold',
-  },
-  input: {
-    width: '100%',
-    height: 50,
-    fontSize: 20,
-    fontWeight: 'bold',
-    borderColor: 'gray',
-    borderWidth: 1,
-    paddingLeft: 10,
-    paddingTop: 10,
-    marginTop: 2,
-    marginBottom: 10,
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 4,
-    },
-    shadowOpacity: 0.3,
-    shadowRadius: 3,
-    elevation: 3,
-    backgroundColor: '#fff',
-  },
-  inputFieldFocused: {
-    borderColor: '#2774D5',
-  },
-  errorInputField: {
-    color: 'red',
-    fontSize: 15,
-    marginLeft: 10,
-  },
-  //LOADING MODAL STYLES
-  modalContainer: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-  },
-  modalContent: {
-    backgroundColor: '#21252b',
-    padding: 20,
-    width: 150,
-    height: 150,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  loadingContainer: {
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    flex: 1,
-  },
-  successContainer: {
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  activityIndicator: {
-    transform: [{ scale: 2.5 }],
-    marginVertical: 25,
-  },
-  iconContainer: {
-    width: 100,
-    height: 100,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  icon: {
-    marginBottom: 15,
-  },
-  signUpModalText: {
-    color: '#2D7BDD',
-    textAlign: 'center',
-    marginTop: 20,
-  },
-
-  //TITLE STYLES
-  title: {
-    fontSize: 26,
-    color: '#2D7BDD',
-    fontWeight: 'bold',
-    marginBottom: 30,
-  },
-
-  //SUBTITLE STYLES
-  subTitle: {
-    color: 'black',
-    fontSize: 35,
-    fontWeight: 'bold',
-  },
-  containerSubTitle: {
-    marginTop: 40,
-    marginBottom: 30,
-  },
-
-  // ORIGIN DROPDOWN STYLES
-  pickerStyle: {
-    borderBottomColor: 'grey',
-    borderBottomWidth: 0.5,
-    borderRadius: 5,
-  },
-  pickerItemStyle: {
-    color: 'lightgray',
-  },
-
-  //FLIGHT INFO
-  flightinfoMainContainer: {},
-
-  flightInfoContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    paddingVertical: 10,
-    marginBottom: 10,
-  },
-  alignRight: {
-    alignItems: 'flex-end',
-  },
-  flightInfoTextContainer: {
-    width: '50%',
-    borderStyle: 'solid',
-    borderBottomWidth: 0.5,
-    borderBottomColor: 'grey',
-  },
-  flightInfoIconContainer: {
-    position: 'absolute',
-    top: '60%',
-    left: '50%',
-    transform: [{ translateX: -12.5 }, { translateY: -12.5 }], //Move Y
-  },
-  flightInfoAbbrev: {
-    fontWeight: '700',
-    fontSize: 25,
-    color: '#000',
-  },
-  flightInfoCountry: {
-    marginBottom: 5,
-  },
-  passengerInfo: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  passengerInfoText: {
-    color: '#000',
-  },
-
-  // BOOKING SCREENS
   flightBookingInfo: {
     flex: 0.3,
     maxHeight: flightInfoHeight,
@@ -302,9 +70,26 @@ const styles = StyleSheet.create({
     marginBottom: '15%',
   },
 
-  //WERE ARE YOU NOW STYLES
   whereAreYouContainer: {
     flex: 1,
+  },
+  flightInfoContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    paddingVertical: 10,
+    marginBottom: 10,
+  },
+  flightInfoTextContainer: {
+    width: '50%',
+    borderStyle: 'solid',
+    borderBottomWidth: 0.5,
+    borderBottomColor: 'grey',
+  },
+  flightInfoIconContainer: {
+    position: 'absolute',
+    top: '60%',
+    left: '50%',
+    transform: [{ translateX: -12.5 }, { translateY: -12.5 }], 
   },
 });
