+++ src/components/OriginDropDown/OriginDropDown.tsx
@@ -1,9 +1,10 @@
 import React, { useState } from 'react';
 import { View } from 'react-native';
 import { Picker } from '@react-native-picker/picker';
-import styles from '../appTheme/AppTheme';
-import flightsData from '../data/flightsData.json';
-import { SelectFlightsProps } from '../interfaces/interfaces';
+import flightsData from '../../data/flightsData.json';
+import { SelectFlightsProps } from '../../interfaces/interfaces';
+import { styles } from './styles';
+
 
 
 export const OriginDropDown = ({ onSelectFlight }: SelectFlightsProps) => {
