+++ src/components/OriginDropDown/styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from "react-native";
+
+
+export const styles = StyleSheet.create ({
+
+    pickerStyle: {
+        borderBottomColor: 'grey',
+        borderBottomWidth: 0.5,
+        borderRadius: 5,
+      },
+      pickerItemStyle: {
+        color: 'lightgray',
+      },
+})
