+++ src/components/SubTitle/SubTitle.tsx
@@ -1,7 +1,7 @@
-import React from 'react'
-import { Text, View } from 'react-native'
-import { SubTitleProps } from '../interfaces/interfaces'
-import styles from '../appTheme/AppTheme'
+import React from 'react';
+import { Text, View } from 'react-native';
+import { SubTitleProps } from '../../interfaces/interfaces';
+import styles from './styles';
 
 export const SubTitle = ({ text } : SubTitleProps) => {
   return (
