+++ src/components/SubTitle/styles.ts
@@ -1,0 +1,15 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  subTitle: {
+    color: 'black',
+    fontSize: 35,
+    fontWeight: 'bold',
+  },
+  containerSubTitle: {
+    marginTop: 40,
+    marginBottom: 30,
+  },
+});
+
+export default styles;
