+++ src/components/Title/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+
+export const styles = StyleSheet.create({
+    title: {
+        fontSize: 26,
+        color: '#2D7BDD',
+        fontWeight: 'bold',
+        marginBottom: 30,
+      },
+})
