+++ src/screens/FlightDestinationScreen.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
 import { View } from 'react-native';
-import { SubTitle } from '../components/SubTitle';
 import { StackScreenProps } from '@react-navigation/stack';
 import styles from '../appTheme/AppTheme';
-import FlightInfo from '../components/FlightInfo';
-import { OriginDropDown } from '../components/OriginDropDown';
-import { ButtonPrimary } from '../components/ButtonPrimary';
+import FlightInfo from '../components/FlightInfo/FlightInfo';
+import { OriginDropDown } from '../components/OriginDropDown/OriginDropDown';
+import { ButtonPrimary } from '../components/ButtonPrimary/ButtonPrimary';
 import Icon from 'react-native-vector-icons/Ionicons';
+import { SubTitle } from '../components/SubTitle/SubTitle';
